[clean]joinCheck(グループ参加確認)の切り出し
